Implement flow animation export to GIF images
We modify the flow_utils.iterate function so that it can now save the flow
animation it creates in Rhinoceros to a GIF file. To do this, the user has
just to provide the path of the output gif file in the 'gif_path' argument.

The implementation uses the 'ViewCaptureToFile' command of Rhinoceros to
write separate frames into a temporary directory; the images2gif.py library
is then used to compile the final GIF file from these frames.

This functionality is going to work only if the user has the Pillow (PIL)
package installed in his or her default Python interpreter. In case that
package is missing, the code will not fail but no GIF file will be produced.